+++ architecture-generator/software_template/driver/include/memIO_tlm.h
@@ -31,10 +31,10 @@ struct TlmMemory : Memory {
 
     public:
 
-    void copyToDevice(uint64_t dest_addr, uint8_t const* src, uint32_t size) override {
+    void copyToDevice(uint64_t dest_addr, uint8_t const* src, uint64_t size) override {
         mem_driver_.write((sctlm::tlm_lib::drivers::addr_type)dest_addr, size, src);
     }
-    void copyFromDevice(uint8_t* dest, uint64_t src_addr, uint32_t size) override {
+    void copyFromDevice(uint8_t* dest, uint64_t src_addr, uint64_t size) override {
         mem_driver_.read((sctlm::tlm_lib::drivers::addr_type)src_addr, size, dest);
     }
     void writeReg32(uint64_t addr, uint32_t value) override {
