+++ architecture-generator/software_template/driver/include/tlmMemory.h
@@ -12,10 +12,10 @@ struct TlmMemory : Memory {
     TlmMemory(sctlm::tlm_lib::drivers::memory_interface &mem, sctlm::tlm_lib::drivers::memory_interface &mgmt): mem_driver_(mem), mgmt_driver_(mgmt) {
     }
 
-    void copyToDevice(uint64_t dest_addr, uint8_t const* src, uint32_t size) override {
+    void copyToDevice(uint64_t dest_addr, uint8_t const* src, uint64_t size) override {
         mem_driver_.write((sctlm::tlm_lib::drivers::addr_type)dest_addr, size, src);
     }
-    void copyFromDevice(uint8_t* dest, uint64_t src_addr, uint32_t size) override {
+    void copyFromDevice(uint8_t* dest, uint64_t src_addr, uint64_t size) override {
         mem_driver_.read((sctlm::tlm_lib::drivers::addr_type)src_addr, size, dest);
     }
     void writeReg32(uint64_t addr, uint32_t value) override {
